requirements.txt
psycopg2-binary==2.9.10
pytz==2025.2
